backend/requirements.txt
-i https://pypi.org/simple
attrs==23.1.0; python_version >= '3.7'
autobahn==23.1.2; python_version >= '3.7'
automat==22.10.0
blinker==1.7.0; python_version >= '3.8'
cffi==1.16.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
constantly==23.10.4; python_version >= '3.8'
contourpy==1.1.1; python_version >= '3.8'
cryptography==41.0.7; python_version >= '3.7'
cycler==0.12.1; python_version >= '3.8'
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
flask==3.0.0; python_version >= '3.8'
flask-cors==4.0.0
fonttools==4.47.0; python_version >= '3.8'
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1; python_version < '3.10'
importlib-resources==6.1.1; python_version < '3.10'
incremental==22.10.0
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
kiwisolver==1.4.5; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.7.4; python_version >= '3.8'
numpy==1.24.4; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pillow==10.1.0; python_version >= '3.8'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pymongo==4.6.1; python_version >= '3.7'
pyopenssl==23.3.0
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
roslibpy==1.6.0
service-identity==23.1.0
setuptools==69.0.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
twisted[tls]==23.10.0; python_full_version >= '3.8.0'
txaio==23.1.1; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
zope-interface==6.1; python_version >= '3.7'
